clean_dir to remove_dir_all

+++ src/utils/fs_ext.rs
@@ -64,11 +64,13 @@ pub async fn make_hard_links_in_dir(
     }
 }
 
+#[allow(dead_code)]
 pub async fn clean_dir(dir: impl AsRef<Path>) -> Result<()> {
     let dir = dir.as_ref();
-    create_dir_if_not_exists(dir).await?;
-    fs::remove_dir_all(dir).await?;
-    create_dir_if_not_exists(dir).await?;
+    if dir.exists() {
+        fs::remove_dir_all(dir).await?;
+        create_dir_if_not_exists(dir).await?;
+    }
     Ok(())
 }
